backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
stripe>=8.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
